% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{basecor_outer}
\alias{basecor_outer}
\title{Outer integrand for base correlation}
\usage{
basecor_outer(v, j, k, copobj, copula, method, type, ...)
}
\arguments{
\item{v}{vector argument of function.}

\item{j}{non-negative integer giving order of first polynomial.}

\item{k}{non-negative integer giving order of second polynomial.}

\item{copobj}{logical parameter for copula object.}

\item{copula}{an object of class parCopula or a function.}

\item{method}{method of calculation which can be "p" or "d".}

\item{type}{type of base correlation can be legendre or cosine.}
}
\value{
value of outer integrand
}
\description{
Outer integrand for base correlation
}
\keyword{internal}
