% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfunctions.R
\name{basiscormatrix}
\alias{basiscormatrix}
\title{Compute Matrix of Basis Correlations}
\usage{
basiscormatrix(object, maxorder = 4, symmetric = FALSE, ...)
}
\arguments{
\item{object}{an object of class copula or Copula, or a data matrix, or a bivariate function describing a copula.}

\item{maxorder}{maximum order of the polynomials.}

\item{symmetric}{logical variable stating whether matrix is known a priori to be symmetric.}

\item{...}{other parameters passed to underlying functions.}
}
\value{
a square matrix with number of rows and columns equal to maxorder
}
\description{
Compute Matrix of Basis Correlations
}
\examples{
basiscormatrix(copula::claytonCopula(2))
}
