% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basisderiv}
\alias{basisderiv}
\title{Derivative of Orthonormal Basis Function}
\usage{
basisderiv(x, degree, type = "legendre")
}
\arguments{
\item{x}{vector of values at which basis function should be evaluated.}

\item{degree}{non-negative integer giving degree of function.}

\item{type}{character string specifying type of basis function and taking values "legendre" or "cosine".}
}
\value{
vector of values of derivative of basis function.
}
\description{
The derivative of \code{\link{basisfunc}}. Used by
\code{\link{basiscorcopula}}'s `method = "p"` integrand.
}
\examples{
basisderiv(seq(0, 1, by = 0.25), 3)
}
