% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{basisexpand}
\alias{basisexpand}
\title{Find Basis Expansion Maximizing Generalized Spearman Correlation}
\usage{
basisexpand(data, maxorder = 8, type = "legendre")
}
\arguments{
\item{data}{matrix of bivariate data}

\item{maxorder}{order of square matrix of basis correlations to be used in expansion}

\item{type}{character variable specifying type of basis correlation taking values "legendre" or "cosine"}
}
\value{
a bex object comprising 4 components: type of basis correlation, maximum value of 
generalized Spearman correlation, weights for first function, and weights for second function.
}
\description{
Find Basis Expansion Maximizing Generalized Spearman Correlation
}
\examples{
data <- copula::rCopula(1000, copula::tCopula(0.5, df = 2))
basisexpand(data, maxorder = 8, type = "legendre")
}
