% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{basisfunc}
\alias{basisfunc}
\title{Orthonormal basis functions}
\usage{
basisfunc(x, degree, type = "legendre")
}
\arguments{
\item{x}{vector of values at which basis function should be evaluated.}

\item{degree}{non-negative integer giving degree of function.}

\item{type}{character string specifying type of basis function and taking values "legendre" or "cosine".}
}
\value{
vector of values of basis function.
}
\description{
Orthonormal basis functions
}
