% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{bexfunc}
\alias{bexfunc}
\title{Evaluate Basis Expansion Function}
\usage{
bexfunc(u, expansion, margin)
}
\arguments{
\item{u}{vector of values at which to compute expansion}

\item{expansion}{a basis expansion given by basisexpand}

\item{margin}{number (1 or 2) specifying the first of second basis expansion function}
}
\value{
vector of values of the function
}
\description{
Evaluate Basis Expansion Function
}
\examples{
data <- copula::rCopula(1000, copula::tCopula(0, df = 1))
bes <- basisexpand(data, maxorder = 8, type = "legendre")
bexfunc(c( 0.2,0.4,0.6), bes, 1)
}
