% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{bexfunc_uni}
\alias{bexfunc_uni}
\title{Evaluate Single Basis Expansion Function}
\usage{
bexfunc_uni(u, alpha, type, k = 0)
}
\arguments{
\item{u}{vector of values at which to compute expansion}

\item{alpha}{vector of weights}

\item{type}{character string specifying type of basis functions, "legendre" or "cosine"}

\item{k}{additional argument for root finding}
}
\description{
Evaluate Single Basis Expansion Function
}
\keyword{internal}
