% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cosine_roots}
\alias{cosine_roots}
\title{Find Roots of Cosine Basis Function}
\usage{
cosine_roots(x, degree = 1)
}
\arguments{
\item{x}{value to find roots for}

\item{degree}{degree of function}
}
\value{
vector of roots
}
\description{
Find Roots of Cosine Basis Function
}
\keyword{internal}
