% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{distLegendre}
\alias{distLegendre}
\title{Distribution of Shifted Legendre Polynomial Transformation of Uniform}
\usage{
distLegendre(x, degree, roottol = 1e-10)
}
\arguments{
\item{x}{vector of values}

\item{degree}{degree of polynomial}

\item{roottol}{tolerance for assigning real roots}
}
\value{
list with two vector components containing values of cumulative distribution function and values of density
}
\description{
Distribution of Shifted Legendre Polynomial Transformation of Uniform
}
\examples{
distLegendre(0.7, degree = 4)
}
