% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dLegendre}
\alias{dLegendre}
\title{Density of Legendre polynomial transformation of uniform}
\usage{
dLegendre(x, degree, tol = 1e-10)
}
\arguments{
\item{x}{vector of values}

\item{degree}{degree of polynomial}

\item{tol}{tolerance for assigning real roots}
}
\value{
values of density
}
\description{
Density of Legendre polynomial transformation of uniform
}
\examples{
dLegendre(0.7, degree = 4)
}
