% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extremalLegendre}
\alias{extremalLegendre}
\title{Calculate extremal Legendre correlation}
\usage{
extremalLegendre(j, k, case = "max")
}
\arguments{
\item{j}{degree of first polynomial}

\item{k}{degree of second polynomial}

\item{case}{character variable which should be max or min}
}
\value{
value of extremal Legendre correlation
}
\description{
Calculate extremal Legendre correlation
}
\examples{
extremalLegendre(3, 4)
}
