% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{legendre_initial}
\alias{legendre_initial}
\title{Collect Data for Distribution of Shifted Legendre Polynomial}
\usage{
legendre_initial(degree)
}
\arguments{
\item{degree}{of polynomial}
}
\value{
list with coefficients of polynomial, coefficients of derivative and lower bound of polynomial
}
\description{
Collect Data for Distribution of Shifted Legendre Polynomial
}
\keyword{internal}
