% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{legendre_roots}
\alias{legendre_roots}
\title{Find Real Roots of Shifted Legendre Polynomial}
\usage{
legendre_roots(legcfs, x, roottol)
}
\arguments{
\item{legcfs}{coefficients of shifted Legendre polynomial}

\item{x}{real value in range of polynomial}

\item{roottol}{tolerance for identification of real roots}
}
\value{
vector of real roots
}
\description{
Find Real Roots of Shifted Legendre Polynomial
}
\keyword{internal}
