% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{pbex}
\alias{pbex}
\title{Evaluate Distribution Function of Basis Expansion of Uniform}
\usage{
pbex(x, expansion, margin, n_grid = 500, tol = 1e-08, h = 1e-06)
}
\arguments{
\item{x}{vector of values at which to compute distribution function}

\item{expansion}{a basis expansion given by basisexpand}

\item{margin}{number (1 or 2) specifying the first or second basis expansion function}

\item{n_grid}{number of grid points for finding turning points and roots of basis expansion function}

\item{tol}{tolerance for finding roots}

\item{h}{tolernace for finding numerical derivatives}
}
\value{
vector of values of the function
}
\description{
Evaluate Distribution Function of Basis Expansion of Uniform
}
\examples{
data <- copula::rCopula(1000, copula::tCopula(0, df = 1))
bes <- basisexpand(data, maxorder = 8, type = "legendre")
pbex(c(-0.5, 4), bes, 1)
}
