% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{pcsm_analysis}
\alias{pcsm_analysis}
\title{Analysis of Piecewise Strictly Monotonic Function}
\usage{
pcsm_analysis(alpha, type, n_grid = 500, plotfunc = FALSE)
}
\arguments{
\item{alpha}{vector of weights}

\item{type}{character string specifying type of basis functions, "legendre" or "cosine"}

\item{n_grid}{number of grid points for finding turning points}

\item{plotfunc}{logical variable specifying whether a plot should be made

keywords internal}
}
\description{
Analysis of Piecewise Strictly Monotonic Function
}
