% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pLegendre}
\alias{pLegendre}
\title{Distribution function of Legendre polynomial transformation of uniform}
\usage{
pLegendre(x, degree, tol = 1e-10)
}
\arguments{
\item{x}{vector of values}

\item{degree}{degree of polynomial}

\item{tol}{tolerance for assigning real roots}
}
\value{
values of cumulative distribution function
}
\description{
Distribution function of Legendre polynomial transformation of uniform
}
\examples{
pLegendre(0.7, degree = 4)
}
