% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{plot.bex}
\alias{plot.bex}
\title{Plot Basis Expansion}
\usage{
\method{plot}{bex}(x, n_grid = 200, udp = FALSE, ...)
}
\arguments{
\item{x}{a basis expansion object created by basisexpand}

\item{n_grid}{number of grid points for plotting function}

\item{udp}{logical value specifying whether to plot udp functions rather than actual basis expansion}

\item{...}{other parameters
functions}
}
\description{
Plot Basis Expansion
}
