% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{polycordata}
\alias{polycordata}
\title{Compute Sample Polynomial Rank Correlation}
\usage{
polycordata(data, j, k, type = "T1")
}
\arguments{
\item{data}{a matrix of data wit two columns.}

\item{j}{non-negative integer giving order of first polynomial.}

\item{k}{non-negative integer giving order of second polynomial.}

\item{type}{method of calculation}
}
\value{
sample polynomial rank correlation value.
}
\description{
Compute Sample Polynomial Rank Correlation
}
