% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{qLegendre}
\alias{qLegendre}
\title{Quantile Function of Shifted Legendre Polynomial Transformation of Uniform}
\usage{
qLegendre(y, degree, roottol = 1e-10, endtol = 1e-10)
}
\arguments{
\item{y}{vector of values}

\item{degree}{degree of polynomial}

\item{roottol}{tolerance for roots of shifted Legendre polynomial}

\item{endtol}{tolerance at endpoints}
}
\value{
values of quantiles
}
\description{
Quantile Function of Shifted Legendre Polynomial Transformation of Uniform
}
\examples{
qLegendre(0.7, degree = 4)
}
