% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rextremalLegendre}
\alias{rextremalLegendre}
\title{Sample from copula attaining extremal Legendre correlation}
\usage{
rextremalLegendre(j, k, n = 1000, case = "max")
}
\arguments{
\item{j}{degree of first polynomial}

\item{k}{degree of second polynomial}

\item{n}{number of sample points}

\item{case}{character variable which should be max or min}
}
\value{
a matrix with nrows and 2 columns containing simulated values from copula
}
\description{
Sample from copula attaining extremal Legendre correlation
}
\examples{
rextremalLegendre(4, 4)
}
