% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{root_analysis}
\alias{root_analysis}
\title{Analysis of Roots of Piecewise Strictly Monotonic Function}
\usage{
root_analysis(k, alpha, type, n_grid = 500, tol = 1e-08, h = 1e-06)
}
\arguments{
\item{k}{real scalar value to find roots for}

\item{alpha}{vector of weights}

\item{type}{character string specifying type of basis functions, "legendre" or "cosine"}

\item{n_grid}{number of grid points for finding turning points}

\item{tol}{tolerance for finding roots}

\item{h}{tolerance for finding numerical derivatives

keywords internal}
}
\description{
Analysis of Roots of Piecewise Strictly Monotonic Function
}
