% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{rstochinv}
\alias{rstochinv}
\title{Bivariate Stochastic Inversion Using Correlation Basis}
\usage{
rstochinv(
  data,
  j,
  k,
  rand = NA,
  type = "legendre",
  roottol = 1e-10,
  endtol = 1e-10
)
}
\arguments{
\item{data}{bivariate matrix of copula data}

\item{j}{degree of first polynomial}

\item{k}{degree of second polynomial}

\item{rand}{randomizer data}

\item{type}{character string giving type of basis correlation}

\item{roottol}{tolerance for roots of shifted Legendre polynomial}

\item{endtol}{tolerance at endpoints for qLegendre}
}
\value{
a matrix with nrows and 2 columns containing simulated values from copula
}
\description{
Bivariate Stochastic Inversion Using Correlation Basis
}
