% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{rstochinvbounds}
\alias{rstochinvbounds}
\title{Sample From Copula Attaining Basis Correlation Bounds}
\usage{
rstochinvbounds(
  j,
  k,
  n = 1000,
  type = "legendre",
  case = "max",
  roottol = 1e-10,
  endtol = 1e-10
)
}
\arguments{
\item{j}{degree of first polynomial}

\item{k}{degree of second polynomial}

\item{n}{number of sample points}

\item{type}{characterstring giving type of basis correlation}

\item{case}{character variable which should be max or min}

\item{roottol}{tolerance for roots of shifted Legendre polynomial}

\item{endtol}{tolerance at endpoints for qLegendre}
}
\value{
a matrix with nrows and 2 columns containing simulated values from copula
}
\description{
Sample From Copula Attaining Basis Correlation Bounds
}
\examples{
rstochinvbounds(4, 4)
}
