% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sLegendre}
\alias{sLegendre}
\title{Shifted Legendre Polynomial}
\usage{
sLegendre(x, degree = 1)
}
\arguments{
\item{x}{vector of values at which polynomial should be evaluated.}

\item{degree}{non-negative integer giving degree of polynomial.}
}
\value{
vector of values of polynomial.
}
\description{
Shifted Legendre Polynomial
}
\examples{
sLegendre(0.5, 2)
}
