% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sLegendreD}
\alias{sLegendreD}
\title{Derivative of Shifted Legendre Polynomial}
\usage{
sLegendreD(x, degree = 1)
}
\arguments{
\item{x}{vector of values at which polynomial should be evaluated.}

\item{degree}{non-negative integer giving degree of polynomial.}
}
\value{
vector of values of polynomial.
}
\description{
The function with non-negative degree n is the derivative of the shifted Legendre of degree n + 1.
}
\examples{
sLegendreD(0.5, 2)
}
