% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sLegendreI}
\alias{sLegendreI}
\title{Integral of Shifted Legendre Polynomial}
\usage{
sLegendreI(x, degree)
}
\arguments{
\item{x}{vector of values at which polynomial should be evaluated.}

\item{degree}{non-negative integer giving degree of polynomial.}
}
\value{
vector of values of integrated polynomial.
}
\description{
Integral of Shifted Legendre Polynomial
}
\examples{
sLegendreI(0.5, 2)
}
