% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{udpbex}
\alias{udpbex}
\title{Evaluate Udp Function for Basis Expansion}
\usage{
udpbex(u, expansion, margin, n_grid = 500, tol = 1e-08, h = 1e-06)
}
\arguments{
\item{u}{vector of values at which to compute udp function}

\item{expansion}{a basis expansion given by basisexpand}

\item{margin}{number (1 or 2) specifying the first or second basis expansion function}

\item{n_grid}{number of grid points for finding turning points and roots of basis expansion function}

\item{tol}{tolerance for finding roots}

\item{h}{tolerance for finding numerical derivatives}
}
\value{
vector of values of the function
}
\description{
Evaluate Udp Function for Basis Expansion
}
\examples{
data <- copula::rCopula(1000, copula::tCopula(0, df = 1))
bes <- basisexpand(data, maxorder = 8, type = "legendre")
udpbex(c(0.2, 0.4, 0.6), bes, 1)
}
