% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansion.R
\name{udpbexdata}
\alias{udpbexdata}
\title{Transform Data with Udp Function for Basis Expansion}
\usage{
udpbexdata(
  Udata,
  expansion,
  boundaryadjust = TRUE,
  btol = NA,
  n_grid = 500,
  tol = 1e-08,
  h = 1e-06
)
}
\arguments{
\item{Udata}{matrix of data with values in (0,1)}

\item{expansion}{a basis expansion given by basisexpand}

\item{boundaryadjust}{logical variable specifying whether to keep transformed data strictly
in (0,1)}

\item{btol}{user-specified adjustment for boundaryadjust}

\item{n_grid}{number of grid points for finding turning points and roots of basis expansion function}

\item{tol}{tolerance for finding roots}

\item{h}{tolerance for finding numerical derivatives}
}
\value{
vector of values of the function
}
\description{
Transform Data with Udp Function for Basis Expansion
}
