% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{udpderiv}
\alias{udpderiv}
\title{Derivative of Uniform Distribution Preserving Transformation for Basis Function}
\usage{
udpderiv(u, degree, type = "legendre", roottol = 1e-10)
}
\arguments{
\item{u}{vector of values between 0 and 1.}

\item{degree}{degree of polynomial.}

\item{type}{character string specifying type of basis function and taking values "legendre" or "cosine".}

\item{roottol}{tolerance passed to distLegendre.}
}
\value{
values of derivative of udp transformation
}
\description{
Derivative of Uniform Distribution Preserving Transformation for Basis Function
}
\examples{
udpderiv(0.3, 3)
}
