% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{udpfunc}
\alias{udpfunc}
\title{Uniform Distribution Preserving Transformation for Basis Function}
\usage{
udpfunc(u, degree, type = "legendre", roottol = 1e-10)
}
\arguments{
\item{u}{vector of values between 0 and 1.}

\item{degree}{degree of polynomial.}

\item{type}{character string specifying type of basis function and taking values "legendre" or "cosine".}

\item{roottol}{tolerance passed to distLegendre.}
}
\value{
values of udp transformation
}
\description{
Uniform Distribution Preserving Transformation for Basis Function
}
\examples{
udpfunc(0.3, 3)
}
