% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{udpLegendre}
\alias{udpLegendre}
\title{Uniform distribution preserving transformation for Legendre polynomial}
\usage{
udpLegendre(u, degree, tol = 1e-10)
}
\arguments{
\item{u}{vector of values between 0 and 1}

\item{degree}{degree of polynomial}

\item{tol}{tolerance passed to pLegendre}
}
\value{
values of udp transformation
}
\description{
Uniform distribution preserving transformation for Legendre polynomial
}
\examples{
udpLegendre(0.3, 3)
}
