% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{udpLegendreD}
\alias{udpLegendreD}
\title{Density of udp transformation for Legendre polynomial}
\usage{
udpLegendreD(u, degree, tol = 1e-10)
}
\arguments{
\item{u}{vector of values between 0 and 1}

\item{degree}{degree of polynomial}

\item{tol}{tolerance passed to dLegendre}
}
\value{
values of density of udp transformation
}
\description{
Density of udp transformation for Legendre polynomial
}
\examples{
udpLegendreD(0.3, 3)
}
