% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{udpLegendreI}
\alias{udpLegendreI}
\title{Data for stochastic inverse of udp Legendre transformation}
\usage{
udpLegendreI(v, degree, tol = 1e-10, data = FALSE, Z = stats::runif(length(v)))
}
\arguments{
\item{v}{vector of values between 0 and 1}

\item{degree}{degree of polynomial}

\item{tol}{tolerance passed to dLegendre and used in root finding}

\item{data}{logical value specifying whether data for stochastic inverse should be returned}

\item{Z}{the randomizer uniform variable which should be same length as v}
}
\value{
either the inverse values or a list consisting of two components - 
a list of possible inverse values and a list of probabilities for each value
}
\description{
Data for stochastic inverse of udp Legendre transformation
}
\examples{
udpLegendre(c(0.4,0.5), degree = 4)
}
