% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{udpstochinv}
\alias{udpstochinv}
\title{Stochastic Inverse for udp Transformation Associated with Basis Function}
\usage{
udpstochinv(v, degree, type = "legendre", roottol = 1e-10, endtol = 1e-10)
}
\arguments{
\item{v}{vector of values between 0 and 1}

\item{degree}{degree of function}

\item{type}{character string specifying type of basis function and taking values "legendre" or "cosine".}

\item{roottol}{tolerance for roots of shifted Legendre polynomial}

\item{endtol}{tolerance at endpoints for qLegendre}
}
\value{
list consisting of two components - 
a list of possible inverse values and a list of probabilities for each value
}
\description{
Stochastic Inverse for udp Transformation Associated with Basis Function
}
\examples{
udpstochinv(c(0.4,0.5), degree = 4)
}
