% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{basiscor}
\alias{basiscor}
\title{Compute Basis Correlation}
\usage{
basiscor(object, j = 1L, k = 1L, ...)
}
\arguments{
\item{object}{an object of class copula or Copula, or a data matrix, or a bivariate function describing a copula.}

\item{j}{non-negative integer giving order of first polynomial.}

\item{k}{non-negative integer giving order of second polynomial.}

\item{...}{other arguments to function.}
}
\value{
value of polynomial rank correlation.
}
\description{
Compute Basis Correlation
}
\examples{
basiscor(copula::claytonCopula(2), 2, 2)
}
