% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfunctions.R
\name{basiscor_inner}
\alias{basiscor_inner}
\title{Inner Integrand for Basis Correlation}
\usage{
basiscor_inner(u, v, j, k, copobj, cop, method, type, ...)
}
\arguments{
\item{u}{vector argument of function.}

\item{v}{vector argument of function.}

\item{j}{non-negative integer giving order of first polynomial.}

\item{k}{non-negative integer giving order of second polynomial.}

\item{copobj}{logical parameter for copula object.}

\item{cop}{an object of class parCopula or a function.}

\item{method}{method of calculation which can be "p" or "d".}

\item{type}{type of basis correlation can be legendre or cosine.}
}
\value{
value of inner integrand
}
\description{
Inner Integrand for Basis Correlation
}
\keyword{internal}
