% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{basiscorcopula}
\alias{basiscorcopula}
\title{Compute Basis Correlation for Copula Object or Function}
\usage{
basiscorcopula(
  copula,
  j = 1L,
  k = 1L,
  copobj,
  method = "p",
  type = "legendre",
  ...
)
}
\arguments{
\item{copula}{an object of class parCopula or a function.}

\item{j}{non-negative integer giving order of first polynomial.}

\item{k}{non-negative integer giving order of second polynomial.}

\item{copobj}{logical parameter for copula object.}

\item{method}{method of calculation which can be "p" or "d".}

\item{type}{type of basis correlation can be legendre or cosine.}

\item{...}{other arguments to function.}
}
\value{
value of polynomial rank correlation.
}
\description{
Compute Basis Correlation for Copula Object or Function
}
\examples{
basiscorcopula(copula::claytonCopula(2), 2, 2, copobj = TRUE)
}
