% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfunctions.R
\name{basiscordata}
\alias{basiscordata}
\title{Compute Sample Basis Correlation}
\usage{
basiscordata(data, j, k, type = "legendre", method = "T1")
}
\arguments{
\item{data}{a matrix of data wit two columns.}

\item{j}{non-negative integer giving order of first polynomial.}

\item{k}{non-negative integer giving order of second polynomial.}

\item{type}{character string specifying type of basis function and taking values "legendre" or "cosine".}

\item{method}{method of calculation}
}
\value{
sample polynomial rank correlation value.
}
\description{
Compute Sample Basis Correlation
}
