% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfunctions.R
\name{discreteLegendre}
\alias{discreteLegendre}
\title{Discrete Legendre Polynomial}
\usage{
discreteLegendre(r, n, degree)
}
\arguments{
\item{r}{rank of observation in sample.}

\item{n}{size of sample/}

\item{degree}{non-negative integer giving degree of polynomial.}
}
\value{
vector of values of polynomial.
}
\description{
Discrete Legendre Polynomial
}
\examples{
discreteLegendre(7, 20, 2)
}
