% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfunctions.R
\name{lff}
\alias{lff}
\title{Log of Falling Factorial Function}
\usage{
lff(r, k)
}
\arguments{
\item{r}{first parameter}

\item{k}{second parameter}
}
\value{
value of function
}
\description{
Log of Falling Factorial Function
}
\keyword{internal}
