% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qLegendre}
\alias{qLegendre}
\title{Quantile function of Legendre polynomial transformation of uniform}
\usage{
qLegendre(y, degree, tol = 1e-10)
}
\arguments{
\item{y}{vector of values}

\item{degree}{degree of polynomial}

\item{tol}{tolerance at endpoints}
}
\value{
values of quantiles
}
\description{
Quantile function of Legendre polynomial transformation of uniform
}
\examples{
qLegendre(0.7, degree = 4)
}
