% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{sLegendre}
\alias{sLegendre}
\title{Shifted Legendre Polynomial and Derivative}
\usage{
sLegendre(x, degree = 1, deriv = FALSE)
}
\arguments{
\item{x}{vector of values at which polynomial should be evaluated.}

\item{degree}{non-negative integer giving degree of polynomial.}

\item{deriv}{logical flag for comuting derivative}
}
\value{
vector of values of polynomial.
}
\description{
Shifted Legendre Polynomial and Derivative
}
\examples{
sLegendre(0.5, 2)
}
